feat: Correctly calculate taxed dividend income
