+++ policyengine_uk/variables/gov/hmrc/income_tax/allowances/allowances.py
@@ -156,20 +156,6 @@ class savings_allowance(Variable):
         )
 
 
-class dividend_allowance(Variable):
-    value_type = float
-    entity = Person
-    label = "Dividend allowance for the person"
-    definition_period = YEAR
-    reference = "Income Tax Act 2007 s. 13A"
-    unit = GBP
-
-    def formula(person, period, parameters):
-        return parameters(
-            period
-        ).gov.hmrc.income_tax.allowances.dividend_allowance
-
-
 class gift_aid(Variable):
     value_type = float
     entity = Person
